feat: add check in LevenbergMarquardtAlgorithm
to see that the empirical pair is compatible with the model function.

We also fall back to previous implementation that is less efficient, but
offers a cleaner interface for the user. In particular, we no longer
compute f-y, but y-f. That means that we can no longer use
subtractInplace, but allows to use theta_k = theta_{k-1} + delta,
instead of theta_k = theta_{k-1} - delta. Then, the user does not need
to know about the implementation details, he knows that the delta
obtained through the algorithm is added to theta_{k-1} to obtain the new
solution.