fix: LevenbergMarquardtAlgorithm so that getErrorLast returns last error
Previously, the error in the previous step was returned. Now we
allocated space for yMinusF, and this is recomputed after updating the
parameters, but used in the next step of the Levenberg Marquardt
algorithm. This makes that getErrorLast returns the error after updating
the parameters through theta_k = theta_{k-1} + delta.