refactor: add ErrorFunction interface
In order to simplify the implementation of the optimization package, we
define a general ErrorFunction. We will drop the definition of multiple
losses that can fit or not to targets. This simplifies the
implementation. In addition, the fact that the error function could be
defined in a tangent space supports this approach.