refactor: loss interfaces
Now the parameters are not set or get. Each class is responsible for
managing and storing the parameters in the most appropriate format. An
external agent will only need to shift the parameters, and not set them.
This approach also allows a cleaner definition of set and get methods
for each specific class.