refactor: redefine Loss interface
Instead of offering to return cost, gradient, or Gauss-Newton matrix, we
now return LossResults. That way we can define a single Loss function
and implement different behaviors depending on the type of results we
ask it for. It is no longer necessary to implement different versions of
a single Loss, nor to implement complex behaviors to avoid
recomputations (for example using the dirty flag optimization pattern).